backend/requirements.txt
# Backend Dependencies
# Phase II: Todo Full-Stack Web Application

# Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# ORM and Database
sqlmodel>=0.0.14
psycopg2-binary>=2.9.9

# Authentication
pyjwt>=2.8.0

# Configuration
python-dotenv>=1.0.0

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Testing (optional)
pytest>=7.4.0
httpx>=0.25.0
